appletb: First cut at a basic touchbar driver.
It turns on the touchbar, and switches it between the special and
function keys when the FN key is pressed or released. In addition it
translates the raw function keys to special keys whenever in special
key mode.

Lastly, it turns the touchbar off if no input (from keyboard, touchpad,
or touchbar) is received for 60 seconds (configurable via module param).