applespi: Switch to using spi-properties exposed by kernel.
The spi_properties work in the kernel now causes the spi device to be
created even for the unmodified DSDT. Furthermore, all properties from
the _DSM method are directly available as acpi device properties, so we
do not need to call and parse that method ourselves anymore.

This reverts commit c13522de3b45fe8e8bb7b07bc774032e6ac187eb as well as
various other pieces.