applespi: Clean up delays.
This is an attempt to fix the last remaining occasional failures due to
insufficient delays. To summarize we now have only two delays:

- the spiCSDelay after asserting CS (for both reads and writes)
- a fixed, 100µs delay whenever switching between writes and reads

In more detail:

spiCSDelay appears to be truly what it says, a delay after asserting CS
and before clocking any data. This also implies that it appears to be in
units of µs. Furthermore this is needed on writes, not just on reads (so
this change includes the write equivalent of commit 3d20af5728f8).

The delay at the end of messages (i.e. before de-asserting CS) appears
to be unnecessary, and was probably an artifact of earlier versions of
this code. Hence removed them.

The resetXXX properties are for resetting the device only, something we
do not currently implement. So stop using them.

Lastly, the 100µs delay between writing and reading is also needed when
switching from reading to writing (otherwise we get occasional (b3 b3 b3
b3) write statuses). Since there is no property in the DSDT that
directly corresponds to this, it is hardcoded for now.