apple-ibridge: moved from MFD to HID driver with virtual HID devices.
This creates one virtual HID device per real HID device and driver pair,
for a total of 4 virtual HID devices. The advantage is that the sub
drivers look very much like as if the drivers were connected to the real
HID devices - e.g. in particular the touch bar driver is aware of the
fact that it is dealing with two HID devices that need to managed
differently.

Signed-off-by: Ronald Tschalär <ronald@innovation.ch>